OC-10758 Provide an ability to inform user when system has failed to send invitation email or message to the participant - Use OCUserDTO
