+++ web/src/main/resources/org/akaza/openclinica/i18n/notes.properties
@@ -784,9 +784,11 @@ upload_instructions=Upload pipe delimited text files which contains CRF data and
 read_log_instructions = Here list all log files about BULK process like import data or schedule event, you can Download the log file to check detail process or Delete it if you don't need any more 
 upload_side_bar_instructions=Batch upload CRF data using a pipe delimited text file. Note: Participants must be created and Study Events must be scheduled prior to importing.
 
-mail_invite_success = Success!<BR><BR>Email invitation to the participant has been sent.
+email_invite_success = Success!<BR><BR>Email invitation to the participant has been sent.
 sms_invite_success = Success!<BR><BR>Mobile invitation to the participant has been sent.
 email_success_sms_fail = Email invitation to the participant has been sent.<BR><BR>Mobile invitation to the participant could not be sent. Please ensure mobile number is correct, or try again later.
 email_fail_sms_success = Mobile invitation to the participant has been sent.<BR><BR>Email invitation to the participant could not be sent. Please ensure email address is correct, or try again later.
+email_invite_fail = Oops!<BR><BR>Email invitation to the participant could not be sent. Please ensure email address, or try again later.
+sms_invite_fail = Oops!<BR><BR>Mobile invitation to the participant could not be sent. Please ensure mobile number is correct, or try again later.
 both_invite_fail = Oops!<BR><BR>Invitation to the participant could not be sent. Please ensure email address and mobile number are correct, or try again later.
 both_invite_success = "Success!<BR><BR>Email and mobile invitations to the participant have been sent.
